Refactor rename
